+++ src/core/field_types/core_double_type.cpp
@@ -3,4 +3,4 @@
 
 using namespace rive;
 
-double CoreDoubleType::deserialize(BinaryReader& reader) { return reader.readFloat32(); }
+float CoreDoubleType::deserialize(BinaryReader& reader) { return reader.readFloat32(); }
